+++ shakki/hashData.h
@@ -8,10 +8,10 @@ class HashData
 {
 public:
 	HashData();
-	HashData( int syvyys, MinMaxPaluu parasSiirto, int vari);
+	HashData( short int syvyys, MinMaxPaluu parasSiirto, short int vari);
 	~HashData();
 
-	int _syvyys;
+	short int _syvyys;
 	MinMaxPaluu _parasSiirto;
-	int _vari;
+	short int _vari;
 };
