+++ shakki/nappula.h
@@ -25,67 +25,67 @@ private:
 	double			_arvo;
 
 public:
-	Nappula(std::wstring, int, int, int);
+	Nappula(std::wstring, short int, short  int, short  int);
 	Nappula() { }
 
 	// Siirtojen generointi. Puhdas virtuaalifunktio, eli aliluokat toteuttavat t‰m‰n
 	// omalla tavallaan.
-	virtual void annaSiirrot(std::list<Siirto>& lista, Ruutu*, Asema*, int vari) = 0;
+	virtual void annaSiirrot(std::list<Siirto>& lista, Ruutu*, Asema*, short int vari) = 0;
 
 	void setUnicode(std::wstring unicode) { _unicode = unicode; }
 	std::wstring getUnicode() { return _unicode; }
-	void setVari(int vari) { _vari = vari; }
-	int getVari() { return _vari; }
-	int getKoodi() { return _koodi; }
+	void setVari(short int vari) { _vari = vari; }
+	short int getVari() { return _vari; }
+	short int getKoodi() { return _koodi; }
 	double getArvo() { return _arvo; }
-	void setKoodi(int koodi) { _koodi = koodi; }
+	void setKoodi(short int koodi) { _koodi = koodi; }
 };
 
 // Torni-aliluokka. Virtuaalinen perint‰ tarkoittaa, ett‰ kantaluokka perit‰‰n moniperinn‰ss‰ vain kerran
 // (koska daami perii sek‰ tornin ett‰ l‰hetin).
 class Torni : public virtual Nappula {
 public:
-	Torni(std::wstring unicode, int vari, int koodi, int arvo) : Nappula(unicode, vari, koodi, arvo) {}
-	void annaSiirrot(std::list<Siirto>& lista, Ruutu*, Asema*, int vari);
+	Torni(std::wstring unicode, short int vari, short int koodi, short int arvo) : Nappula(unicode, vari, koodi, arvo) {}
+	void annaSiirrot(std::list<Siirto>& lista, Ruutu*, Asema*, short int vari);
 };
 
 // Ratsu-aliluokka.
 class Ratsu : public Nappula {
 public:
-	Ratsu(std::wstring unicode, int vari, int koodi, int arvo) : Nappula(unicode, vari, koodi, arvo) {}
-	void annaSiirrot(std::list<Siirto>& lista, Ruutu*, Asema*, int vari);
+	Ratsu(std::wstring unicode, short int vari, short int koodi, short int arvo) : Nappula(unicode, vari, koodi, arvo) {}
+	void annaSiirrot(std::list<Siirto>& lista, Ruutu*, Asema*, short int vari);
 };
 
 // L‰hetti-aliluokka. Virtuaalinen perint‰ tarkoittaa, ett‰ kantaluokka perit‰‰n moniperinn‰ss‰ vain kerran
 // (koska daami perii sek‰ tornin ett‰ l‰hetin).
 class Lahetti : public virtual Nappula {
 public:
-	Lahetti(std::wstring unicode, int vari, int koodi, int arvo) : Nappula(unicode, vari, koodi, arvo) {}
-	void annaSiirrot(std::list<Siirto>& lista, Ruutu*, Asema*, int vari);
+	Lahetti(std::wstring unicode, short int vari, short int koodi, short int arvo) : Nappula(unicode, vari, koodi, arvo) {}
+	void annaSiirrot(std::list<Siirto>& lista, Ruutu*, Asema*, short int vari);
 };
 
 // Daami-aliluokka. Perii sek‰ l‰hetin ett‰ tornin.
 class Daami : public Lahetti, public Torni {
 public:
-	Daami(std::wstring unicode, int vari, int koodi, int arvo) :
+	Daami(std::wstring unicode, short int vari, short int koodi, short int arvo) :
 		Nappula(unicode, vari, koodi, arvo), Lahetti(unicode, vari, koodi, arvo), Torni(unicode, vari, koodi, arvo) {}
-	void annaSiirrot(std::list<Siirto>& lista, Ruutu*, Asema*, int vari);
+	void annaSiirrot(std::list<Siirto>& lista, Ruutu*, Asema*, short int vari);
 };
 
 // Kuningas-aliluokka.
 class Kuningas : public Nappula {
 public:
-	Kuningas(std::wstring unicode, int vari, int koodi, int arvo) : Nappula(unicode, vari, koodi, arvo) {}
-	void annaSiirrot(std::list<Siirto>& lista, Ruutu*, Asema*, int vari);
+	Kuningas(std::wstring unicode, short int vari, short int koodi, short int arvo) : Nappula(unicode, vari, koodi, arvo) {}
+	void annaSiirrot(std::list<Siirto>& lista, Ruutu*, Asema*, short int vari);
 };
 
 // Sotilas-aliluokka.
 class Sotilas : public Nappula {
 public:
-	Sotilas(std::wstring unicode, int vari, int koodi, int arvo) : Nappula(unicode, vari, koodi, arvo) {}
-	void annaSiirrot(std::list<Siirto>& lista, Ruutu*, Asema*, int vari);
+	Sotilas(std::wstring unicode, short int vari, short int koodi, short int arvo) : Nappula(unicode, vari, koodi, arvo) {}
+	void annaSiirrot(std::list<Siirto>& lista, Ruutu*, Asema*, short int vari);
 private:
-	void lisaaSotilaanKorotukset(Siirto, std::list<Siirto>& lista, int vari);
+	void lisaaSotilaanKorotukset(Siirto, std::list<Siirto>& lista, short int vari);
 };
